Own files for i18n and router + fix some of markdown warnings
