+++ CONTRIBUTING.md
@@ -2,7 +2,7 @@
 
 First of all, thank you everyone who made pull requests for Uptime Kuma, I never thought GitHub Community can be that nice! And also because of this, I also never thought other people actually read my code and edit my code. It is not structed and commented so well, lol. Sorry about that.
 
-The project was created with vite.js (vue3). Then I created a sub-directory called "server" for server part. Both frontend and backend share the same package.json. 
+The project was created with vite.js (vue3). Then I created a sub-directory called "server" for server part. Both frontend and backend share the same package.json.
 
 The frontend code build into "dist" directory. The server uses "dist" as root. This is how production is working.
 
@@ -20,13 +20,13 @@ If you are not sure, feel free to create an empty pull request draft first.
 - Add a chart
 - Fix a bug
 
-### *️⃣ Requires one more reviewer 
+### *️⃣ Requires one more reviewer
 
 I do not have such knowledge to test it.
 
-- Add k8s supports 
+- Add k8s supports
 
-### *️⃣ Low Priority 
+### *️⃣ Low Priority
 
 It changed my current workflow and require further studies.
 
@@ -41,9 +41,9 @@ It changed my current workflow and require further studies.
 
 # Project Styles
 
-I personally do not like something need to learn so much and need to config so much before you can finally start the app. 
+I personally do not like something need to learn so much and need to config so much before you can finally start the app.
 
-For example, recently, because I am not a python expert, I spent a 2 hours to resolve all problems in order to install and use the Apprise cli. Apprise requires so many hidden requirements, I have to figure out myself how to solve the problems by Google search for my OS. That is painful. I do not want Uptime Kuma to be like this way, so:  
+For example, recently, because I am not a python expert, I spent a 2 hours to resolve all problems in order to install and use the Apprise cli. Apprise requires so many hidden requirements, I have to figure out myself how to solve the problems by Google search for my OS. That is painful. I do not want Uptime Kuma to be like this way, so:
 
 - Easy to install for non-Docker users, no native build dependency is needed (at least for x86_64), no extra config, no extra effort to get it run
 - Single container for Docker users, no very complex docker-composer file. Just map the volume and expose the port, then good to go
@@ -52,8 +52,8 @@ For example, recently, because I am not a python expert, I spent a 2 hours to re
 
 # Coding Styles
 
-- Follow .editorconfig
-- Follow eslint
+- Follow `.editorconfig`
+- Follow ESLint
 
 ## Name convention
 
@@ -62,12 +62,13 @@ For example, recently, because I am not a python expert, I spent a 2 hours to re
 - CSS/SCSS: dash-type
 
 # Tools
+
 - Node.js >= 14
 - Git
-- IDE that supports .editorconfig and eslint (I am using Intellji Idea)
+- IDE that supports EditorConfig and ESLint (I am using Intellji Idea)
 - A SQLite tool (I am using SQLite Expert Personal)
 
-# Install dependencies 
+# Install dependencies
 
 ```bash
 npm install --dev
@@ -75,7 +76,7 @@ npm install --dev
 
 For npm@7, you need --legacy-peer-deps
 
-```
+```bash
 npm install --legacy-peer-deps --dev
 ```
 
@@ -84,23 +85,22 @@ npm install --legacy-peer-deps --dev
 ```bash
 npm run start-server
 
-# Or 
+# Or
 
 node server/server.js
 ```
 
-It binds to 0.0.0.0:3001 by default.
-
+It binds to `0.0.0.0:3001` by default.
 
 ## Backend Details
 
 It is mainly a socket.io app + express.js.
 
-express.js is just used for serving the frontend built files (index.html, .js and .css etc.) 
+express.js is just used for serving the frontend built files (index.html, .js and .css etc.)
 
 # Frontend Dev
 
-Start frontend dev server. Hot-reload enabled in this way. It binds to 0.0.0.0:3000.
+Start frontend dev server. Hot-reload enabled in this way. It binds to `0.0.0.0:3000` by default.
 
 ```bash
 npm run dev
@@ -108,7 +108,7 @@ npm run dev
 
 PS: You can ignore those scss warnings, those warnings are from Bootstrap that I cannot fix.
 
-You can use Vue Devtool Chrome extension for debugging.
+You can use Vue.js devtools Chrome extension for debugging.
 
 After the frontend server started. It cannot connect to the websocket server even you have started the server. You need to tell the frontend that is a dev env by running this in DevTool console and refresh:
 
@@ -118,8 +118,7 @@ localStorage.dev = "dev";
 
 So that the frontend will try to connect websocket server in 3001.
 
-Alternately, you can specific NODE_ENV to "development".
-
+Alternately, you can specific `NODE_ENV` to "development".
 
 ## Build the frontend
 
@@ -131,22 +130,17 @@ npm run build
 
 Uptime Kuma Frontend is a single page application (SPA). Most paths are handled by Vue Router.
 
-The router in "src/main.js"
+The router is in `src/router.js`
 
 As you can see, most data in frontend is stored in root level, even though you changed the current router to any other pages.
 
-The data and socket logic in "src/mixins/socket.js"
+The data and socket logic are in `src/mixins/socket.js`.
 
 # Database Migration
 
-1. create `patch{num}.sql` in `./db/`
-1. update `latestVersion` in `./server/database.js`
+1. Create `patch{num}.sql` in `./db/`
+2. Update `latestVersion` in `./server/database.js`
 
 # Unit Test
 
 Yes, no unit test for now. I know it is very important, but at the same time my spare time is very limited. I want to implement my ideas first. I will go back to this in some points.
-
-
-
-  
-
